separate method visitors for adapters and input validators
