+++ dd-java-agent/agent-iast/src/main/java/com/datadog/iast/securitycontrol/SecurityControlMethodClassVisitor.java
@@ -1,17 +1,22 @@
 package com.datadog.iast.securitycontrol;
 
-import static org.objectweb.asm.Opcodes.ACC_STATIC;
 import static org.objectweb.asm.Opcodes.ASM8;
 
 import datadog.trace.api.iast.securitycontrol.SecurityControl;
+import datadog.trace.api.iast.securitycontrol.SecurityControlType;
 import java.util.List;
 import javax.annotation.Nullable;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.MethodVisitor;
+import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class SecurityControlMethodClassVisitor extends ClassVisitor {
+
+  static final Logger LOGGER = LoggerFactory.getLogger(SecurityControlMethodClassVisitor.class);
 
   private final List<SecurityControl> securityControls;
 
@@ -37,8 +42,12 @@ public class SecurityControlMethodClassVisitor extends ClassVisitor {
       }
     }
     if (match != null) {
-      final boolean isStatic = (access & ACC_STATIC) > 0;
-      return new SecurityControlMethodAdapter(mv, match, desc, isStatic);
+      final Type method = Type.getMethodType(desc);
+      if (match.getType() == SecurityControlType.SANITIZER) {
+        mv = adaptSanitizer(mv, match, access, method);
+      } else {
+        mv = adaptInputValidator(mv, match, access, method);
+      }
     }
     return mv;
   }
@@ -69,5 +78,35 @@ public class SecurityControlMethodClassVisitor extends ClassVisitor {
     }
 
     return true;
+  }
+
+  private MethodVisitor adaptSanitizer(
+      final MethodVisitor mv,
+      final SecurityControl control,
+      final int accessFlags,
+      final Type method) {
+    if (isPrimitive(method.getReturnType())) {
+      // no need to check primitives as we are not tainting them
+      LOGGER.warn(
+          "Sanitizers should not be used on non-primitive return types. Return type {}. Security control: {}",
+          method.getReturnType().getClassName(),
+          control);
+      return mv;
+    }
+    return new SanitizerMethodAdapter(mv, control, accessFlags, method);
+  }
+
+  private MethodVisitor adaptInputValidator(
+      final MethodVisitor mv,
+      final SecurityControl control,
+      final int accessFlags,
+      final Type method) {
+    return new InputValidatorMethodAdapter(mv, control, accessFlags, method);
+  }
+
+  public static boolean isPrimitive(final Type type) {
+    // Check if is a primitive type
+    int sort = type.getSort();
+    return sort >= Type.BOOLEAN && sort <= Type.DOUBLE;
   }
 }
